Add files via upload
Added the updated research_code.py as well as putting in the flux and wavelength information in the spectra.npz file so that i can test out my emission line code, line flux, line equivalent width. I also added in a target .dat files and linelist.dat which are filename information and linelist of interest to us.