Fixed redshift issues and emission lines matched
Now i need to make a table with all the emission line information. Then perform some line ratio analysis and then plot it on the BPT diagram with the rest of the galaxies from SDSS. I also need to perform some statistical stuff for when Micaela gives me the errors on extraction, reddening and subtraction.